Initial commit
Basic classes for using GTK+3 with xojo 2017r2 and newer.
GTKLibWindow has a createHeaderbar event that tries to be a workaround around Xojo Feedback case <feedback://showreport?report_id=50044> but succeeds only in rare cases like Ubuntu. This is why this library is free for experiments helping to fix this case as it currently seems to be s showstopper for further GTK enhancements. Please see the different values for Window and widget height that are shown on the istbox before and after the Headerbar.

Signed-off-by: UBogun <bogun@satzservice.de>